Fix initialization of Ribbit::Commit
The Commit object was not properly inheriting the original
Repository Object; all the common methods are working now,
including the 'read_raw'.

Commit objects, like all the other repository objects, now
follows the new instantation rules:

	- the repository parameter is always required.
	- not supplying a SHA1 ID implies the object
will be created on memory to be later written to the repository.
	- supplying a SHA1 ID will lookup the object on the
repository and instantiate it.

Signed-off-by: Vicent Marti <tanoku@gmail.com>