Fix object writeback and raw reads
The two broken unit tests are now passing.

	- For object raw read, the SHA1 of the object was being obtained
properly from the "@sha" attribute, which was deprecated and replaced
with a "sha" getter method. This method is now properly called.

	- For object write-back, the writing of commits was failing because
when doing a repository lookup for a commit, only the minimal amount of
data was being parsed and loaded (tree, time and parents), and hence the
library didn't have enough data to write back the commit on a
modification. Libgit2 has changed and now properly loads all data before
trying to modify a commit in-memory.

Signed-off-by: Vicent Marti <tanoku@gmail.com>