Iterate and enumerate like it was meant to be
Fixed all the enumerators in the library to return themselves instead of
a proxy to array when they are called without a block.

This includes the Revwalker, which now implements the `each` method for
enumeration. The old `next` method is still available, just call the
walk without a block:

	# `walk` and `each` are alias
	@walker.each { |commit| ... }
	@walker.walk { |commit| ... }

	w = @walker.walk # => Enumerable
	w.next
	w.next

PS: I renamed all instances of `sha` to `oid`. Why? Because I'm a moron.