Change :name key to :path in the README's Builder example
In README.md, :name is described as the right key to use when building a tree.
However, when I pass that key, I get an exception from Builder#<<... something
along the lines of "invalid type nil, expecting String".

Looking at the C code, I see mention of "path" as a key, and passing my
path component in the :path slot fixes the problem.  (Note this is in
contrast to iterating through the entries of a tree object; in that case,
the key used is indeed :name.)

Change the docs to match what's working.