Make native extension building more robust
Changed gemspec to include proper directives for the native extension library so
that it can be found on every platform.

Removed the dependency on having libgit2 built in vendor/bundle and instead
provided --with-libgit2-{dir,include,lib} build options with dir_config in
extconf.rb. If libgit2 is installed system wide it will be found in the normal
directories, if it is installde in a non-standard location the location can be
supplide with the new build options. When desired libgit2 can be built
statically in any location, the new build options can then be used to specify
the static version to be preferred over any system version.

The new build options can be supplied to extconf.rb directly, to gem install
with e.g. gem install rugged -- --with-libgit2-include=vendor/libgit2/include
--with-libgit2-lib=vendor/libgit2/build or to bundle config with e.g. bundle
config rugged.build --with-libgit2-dir=/opt/libgit2