Clean up Rugged::Diff::Patch line stats APIs
This removes the #additions, #deletions, and #context APIs from
the patch object in favor of the #stat API, plus it adds a new
API Rugged::Diff::Patch#lines that yields the total lines in the
whole patch including context.  Using that, the context value is
still available via `patch.lines - patch.changes`

The tests are updated to cover all these APIs.