Ignore library search paths returned by `pkg-config`.
Instead of adding the output of `pkg-config --libs --static` directly
to `$LDFLAGS`, we append only the `-l` part to `$libs`.

This is to fix an issue where trying to build rugged with a
pre-installed libgit2 version would link against that version.

`pkg-config --libs --static` includes something along the lines
of `-L${libdir}`, where `${libdir}` could point to the folder
containing the pre-installed libgit2 version.