Update README [skip ci]
Add `index.read_tree(repo.head.target.tree)` to README, it took me a good hour to find out I need to write the current tree into the index otherwise I would lose the files that are not in the there. Just nice to have that on the README so people wont make the same mistake.