Build on Windows with WinHTTP
Insert the right libraries to link to to get WinHTTP working. These are
normally done by the build system, but we're building statically so we put
them in ourselves.

We don't write out any dependencies to pkg-config, so let's get rid of it.
To boot, it's not always installed on this platform.