commit: Cleanup header extraction APIs
Just some simple changes since I missed review the previous time.
`Commit#header_field?` can be implemented directly in Ruby since the C
code is almost 100% redundant.

The original `Commit#header_field` has been refactored to prevent a
possible memory leak, and to properly tag the resulting header with the
encoding of the commit message, if we know it.