extconf: Use MSYS Makefiles with CMake on Windows
On Windows, Ruby-DevKit is MSYS-based. As such we should use MSYS
Makefiles with CMake. Otherwise it might happen that CMake picks a
gmake.exe from somewhere instead of make.exe from Ruby-DevKit as CMake
seems to prefer gmake over make.

In a concrete case, when doing "gem install rugged" on Windows with
Strawberry Perl installed and in PATH, CMake was picking

    C:/Strawberry/c/bin/gmake.exe

instead of

    C:/Ruby-DevKit/bin/make.exe

While the latter is linked against MSYS-1.0.DLL and understands Unix-style
paths on Windows, the first is a native Windows executable linked against
MSVCRT.DLL that does not. This resulted in path mangling issues leading to
CMake's "Detecting C compiler ABI info" step to fail because the helper
executable failed to build, which in turn led to CMAKE_SIZEOF_VOID_P not
being set and libgit2 being unable to detect the architecture.

On other platforms than Windows it should be safe to use whatever the
default toolchain is that CMake detects.