Test `Index.merge_file` with missing sides
`Index.merge_file` produces the conflict file for a given conflict.
This can be produced for add/add conflicts (that are missing ancestors)
but - by definition - cannot be produced for edit/delete conflicts.

(Git simply leaves the edited file in the working tree without conflict
markup.)

Validate that we handle this case correctly.