Don't default index entries to "VALID"
The "VALID" flag in git refers to the "assume-unchanged" feature, such
that if, in a non-bare repository, an index entry is marked as "VALID",
the git CLI (e.g. `git status`) will assume that the file never changes,
and ignore any changes that do occur.

See https://github.com/libgit2/rugged/issues/636 for the issue and
research leading to this change.