Set a custom allocator to redirect via the ruby runtime
This lets us tell the ruby runtime how much memory rugged/libgit2 is allocating
to give it a more realistic view of how much memory the process is using.

The expectation/hope is that long-lived processes will give back memory to the
system more readily than is currently the case.