Use strnlen in the strndup allocator
If we use `strlen` we'll go over the whole string rather than just the part that
interests us, and sometimes we are given large strings.

As en example, a 15MB reflog took ~5s to parse when using `strlen` and it fell
within the noise of the tests when using `strnlen`.