Give CMake a timeout for when it hangs
In Debian Jessie, CMake will sometimes fail to correctly wait for the programs
it has spawned. While a proper fix would be in CMake itself, protect ourselves
from this by giving it a generous but limited time in which to run.

When it hangs, it will now instead be killed by the build process and we can try
again, instead of having to hunt down processes across machines.