fix broken Windows builds
First issue: pgroup is not available on Windows. new_pgroup should be
used instead. This broke in commit 790755c66a94f3b6088592367598706a97d9a530
(#763). Fixes #791.

Second issue: the mingw libssh2 package, which is required for building rugged,
is not installed out of the box on RubyInstaller with MSYS2.
This PR adds it to the dependency list so that installing the rugged gem
automatically pulls in the mingw libssh2 package as documented here:
https://github.com/oneclick/rubyinstaller2/wiki/For-gem-developers#-msys2-library-dependency
(related to PR #770).