ci: be less precise with our ruby version with setup-ruby
We constantly are fighting updating these minor versions because rbenv wants us
to be precise. But the ruby-setup action we rely on for the Ubuntu builds
has a much more restricted list of versions which are available.

Keeping both as the same list means CI breaks constantly for no reason so create
two different lists to try to stop us having to do busywork.