ci: split into two distinct jobs
The previous commit introduced two lists of ruby to use, but that meant that we
now had triples to test, exploding the amount of tests without reason.

In order to avoid this, split the build into a macOS-specific build and a
Ubuntu-specific one which avoids the branching though it does lead to some
duplicated code.

It does however still achieve the goal of not having to change the Ubuntu
build's ruby versions constantly which is still less overhead.