Add documentation on installing rugged with SSH
As of libgit2 v1.4.0 USE_SSH defaults to off, this results in a behavior
change from the end-user point-of-view in that rugged no longer has the
SSH feature enabled even if libssh2 is installed where this worked on
rugged v1.3.

Document the CMAKE_FLAGS needed to install rugged v1.4 with the SSH
feature enabled.